we can parse all sorts of const expressions now! Probably time to move on to another type of item for a bit.
