+++ Cargo.toml
@@ -6,8 +6,8 @@ version = "0.1.0"
 edition = "2021"
 license = "Zlib OR Apache-2.0 OR MIT"
 
-[profile.dev]
-opt-level=3
+[profile.dev.package."*"]
+opt-level = 3
 
 [profile.release]
 codegen-units=1
