+++ examples/scratch.yag
@@ -1,2 +1,6 @@
-
-const FOO = 0;
+const VRAM_TILE_BLOCK_2 = $9000;
+const VRAM_TILE_MAP_0 = $9800;
+const NR52 = $FF26;
+const LCDC = $FF40;
+const LY = $FF44;
+const BGP = $FF47;
